Creating Medicine Shop Page
Incomplete for Deletion.
No Update yet.

Medicineshop_Page will use Medicine Card that will retrieve medicineListShop collection document from firebase.

User can update Medicine by clicking the floating action button.

Then can click on it to view and if they want to pay it will lead to the payment Page. PaymentPage route is not configured yet.

The only problem is on the delete medicine card. It is by swiping right on the cards with option to Undo the action. But this still output error.

So may update soon