added Update option
Uses the updateMedicine controller functions where it will find the medicine by id in the document collection then update it by mapping the medicine object class to the updated medicine.

The rest is for update medicine page.dart where the void updateMedicine function is being used to update all medicine attributes by declaring updatedMedicine as a MedicalPrescriptionModel class with the textController being used to update it. Then use firestore await syntax to call updatedMedicine function and store updatedMedicine MedicalPrescriptionModel as a parameter to update the medicine