Added CartPage and Change user roles for each MedicineShopPage
Each user roles for 'user' and 'admin' will now have separate view page for MedicineShopPage. This uses the fetchCurrentUserRole Future<void> Function in the MedicineShopPage and  currentUserRole String variable.
For those I added required this.currentUserRole dependancy for MedicineCard and MedicineDetailsPage. So dev can just put 'currentUserRole == currentUserRole' syntax when trying to go to the each respective modules or page so that currentUserRole will be passed along with it as well.

For CartPage, only user can view, add into cart, and remove items from cart. Then user can select View Cart button that will lead the to the CartPage.

For now I haven't make the PaymentPage yet so 'Checkout' Button will not redirect the user to anywhere.
Also, the items listed in CartPage will show the same stacked list for the same medicine, so no functions for one medicine have this specific amount etc.