Update Report so it will fetch BMI result, Height and Weight
BMI Result on the ListTile
Height and Weight on the Cards

Also uses ReportModel, ReportController, and ReportTab.

to show BMI details are properly retrieved when the user logs in, _fetchUserData async will call the reportcontroller and setState to refresh the page. Then InitState will use the _fetchUserData to initialized the page with the BMI details