Updated the PaymentPage.
ListView.Builder does not work but somehow StreamBuilder does to display the medicine.

 Also update the code for PurchaseSummary so it wont repeat the same item more than once with in large quantities.

Also add clearCartItems so it everytime user go back to Cart page, the items in the cart purchase will be removed but not the items for the cart themselves. This prevents the user for purchasing the same item and quantity repeatedly when they go back to the PurchaseSummaryPage.